agents/content_agents/requirements.txt
google-generativeai==0.3.0
tweepy==4.14.0
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.0
apscheduler==3.10.4
